Make the api directory structure uniform
