+++ ui/api/kafka/actions.ts
@@ -1,49 +1,14 @@
+"use server";
 import { getHeaders } from "@/api/api";
+import {
+  ClusterDetail,
+  ClusterList,
+  ClusterResponse,
+  ClustersResponseSchema,
+} from "@/api/kafka/schema";
 import { logger } from "@/utils/logger";
-import { z } from "zod";
 
 const log = logger.child({ module: "kafka-api" });
-
-const NodeSchema = z.object({
-  id: z.number(),
-  host: z.string(),
-  port: z.number(),
-  rack: z.string().optional(),
-});
-export type KafkaNode = z.infer<typeof NodeSchema>;
-
-export const ClusterListSchema = z.object({
-  id: z.string(),
-  type: z.string(),
-  attributes: z.object({
-    name: z.string(),
-    namespace: z.string(),
-    bootstrapServers: z.string(),
-  }),
-});
-export const ClustersResponseSchema = z.object({
-  data: z.array(ClusterListSchema),
-});
-export type ClusterList = z.infer<typeof ClusterListSchema>;
-
-const ClusterDetailSchema = z.object({
-  id: z.string(),
-  type: z.string(),
-  attributes: z.object({
-    name: z.string(),
-    namespace: z.string(),
-    creationTimestamp: z.string(),
-    nodes: z.array(NodeSchema),
-    controller: NodeSchema,
-    authorizedOperations: z.array(z.string()),
-    bootstrapServers: z.string(),
-    authType: z.string().optional().nullable(),
-  }),
-});
-export const ClusterResponse = z.object({
-  data: ClusterDetailSchema,
-});
-export type ClusterDetail = z.infer<typeof ClusterDetailSchema>;
 
 export async function getKafkaClusters(): Promise<ClusterList[]> {
   const url = `${process.env.BACKEND_URL}/api/kafkas?fields%5Bkafkas%5D=name,namespace,bootstrapServers`;
