+++ ui/api/nodes/actions.ts
@@ -1,29 +1,9 @@
+"use server";
 import { getHeaders } from "@/api/api";
+import { ConfigResponseSchema, NodeConfig } from "@/api/nodes/schema";
 import { logger } from "@/utils/logger";
-import { z } from "zod";
 
 const log = logger.child({ module: "api-topics" });
-
-const ConfigSchema = z.object({
-  id: z.string().optional(),
-  type: z.string(),
-  meta: z.record(z.any()).optional(),
-  attributes: z.record(
-    z.object({
-      value: z.string().optional(),
-      source: z.string().readonly(),
-      sensitive: z.boolean().readonly(),
-      readOnly: z.boolean().readonly(),
-      type: z.string().readonly(),
-      documentation: z.string().readonly().optional(),
-    }),
-  ),
-});
-export type NodeConfig = z.infer<typeof ConfigSchema>;
-
-const ConfigResponseSchema = z.object({
-  data: ConfigSchema,
-});
 
 export async function getNodeConfiguration(
   kafkaId: string,
