+++ ui/app/[locale]/kafka/[kafkaId]/@header/KafkaHeader.tsx
@@ -1,4 +1,4 @@
-import { getKafkaCluster } from "@/api/kafka";
+import { getKafkaCluster } from "@/api/kafka/actions";
 import { getTopics } from "@/api/topics/actions";
 import { KafkaParams } from "@/app/[locale]/kafka/[kafkaId]/kafka.params";
 import { AppHeader } from "@/components/AppHeader";
