+++ ui/app/[locale]/kafka/[kafkaId]/@title/KafkaTitle.tsx
@@ -1,4 +1,4 @@
-import { getKafkaCluster } from "@/api/kafka";
+import { getKafkaCluster } from "@/api/kafka/actions";
 import { KafkaParams } from "@/app/[locale]/kafka/[kafkaId]/kafka.params";
 import { Title } from "@/libs/patternfly/react-core";
 import { Skeleton } from "@patternfly/react-core";
