+++ ui/app/[locale]/kafka/[kafkaId]/topics/create/page.tsx
@@ -1,4 +1,4 @@
-import { getKafkaCluster } from "@/api/kafka";
+import { getKafkaCluster } from "@/api/kafka/actions";
 import { createTopic } from "@/api/topics/actions";
 import { NewConfigMap } from "@/api/topics/schema";
 import { KafkaParams } from "@/app/[locale]/kafka/[kafkaId]/kafka.params";
